code/requirements.txt
flask==2.0.1
tensorflow==2.15.0
pillow==9.0.0
pillow-heif==0.7.0
numpy>=1.23.5,<2.0.0
matplotlib>=3.5.0 